Fix-bug: switch.c: Fix oversized write
The maximum size for a zone context should be `MAX_COMMAND`, which was
correctly used by the unit test. But the rest of the application used
the size `MAX_FIELD` for writing to the string. Fix this by using the
correct size for the string.

Additionally, fix several trailing whitespace issues.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>